commit of lesson 1-2 exported child components to App
